feat: improve chat screen (v3)
- reduce number of states of sidebar
- sync sidebar state with agent picker state
